+++ http/src/main/java/net/daporkchop/lib/http/client/HttpClient.java
@@ -15,6 +15,9 @@
 
 package net.daporkchop.lib.http.client;
 
+import io.netty.util.concurrent.Future;
+import net.daporkchop.lib.http.HttpEngine;
+
 /**
  * A container around an instance of an HTTP client.
  * <p>
@@ -25,5 +28,22 @@ package net.daporkchop.lib.http.client;
  * @author DaPorkchop_
  */
 public interface HttpClient {
-    //TODO
+    /**
+     * @return the {@link HttpEngine} that created this {@link HttpClient}
+     */
+    HttpEngine engine();
+
+    /**
+     * Closes this client instance.
+     * <p>
+     * This will release any resources allocated specifically by this client (however, any resources allocated
+     * by the parent {@link HttpEngine} will not be released unless specifically stated otherwise by the
+     * implementation).
+     */
+    Future<Void> close();
+
+    /**
+     * @return a {@link Future} that will be notified when this {@link HttpClient} is closed
+     */
+    Future<Void> closeFuture();
 }
