+++ http/src/main/java/net/daporkchop/lib/http/server/HTTPServer.java
@@ -15,69 +15,31 @@
 
 package net.daporkchop.lib.http.server;
 
-import io.netty.bootstrap.ServerBootstrap;
-import io.netty.buffer.Unpooled;
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.ChannelInboundHandlerAdapter;
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.nio.NioServerSocketChannel;
-import net.daporkchop.lib.common.util.PorkUtil;
-import net.daporkchop.lib.http.Request;
-import net.daporkchop.lib.http.Response;
-import net.daporkchop.lib.http.codec.ExceptionHandlerServerHTTP;
-import net.daporkchop.lib.http.codec.v1.RequestDecoderHTTP1;
-import net.daporkchop.lib.http.codec.v1.ResponseEncoderHTTP1;
-import net.daporkchop.lib.http.util.ConnectionState;
-import net.daporkchop.lib.http.util.StatusCodes;
-import net.daporkchop.lib.http.util.exception.GenericHTTPException;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Collections;
-import java.util.Scanner;
-
-import static net.daporkchop.lib.http.util.Constants.*;
+import io.netty.util.concurrent.Future;
+import net.daporkchop.lib.http.HttpEngine;
 
 /**
- * Test class for things, will probably be turned into an actual HTTP server helper at some point.
+ * The core of an HTTP server implementation.
  *
  * @author DaPorkchop_
  */
-public class HTTPServer {
-    public static void main(String... args) throws Exception {
-        Channel ch = new ServerBootstrap()
-                .group(new NioEventLoopGroup())
-                .channelFactory(NioServerSocketChannel::new)
-                .childHandler(new ChannelInitializer() {
-                    @Override
-                    protected void initChannel(Channel ch) throws Exception {
-                        ch.attr(KEY_STATE).set(ConnectionState.INIT);
-                        ch.pipeline()
-                                .addLast("http_decode", new RequestDecoderHTTP1())
-                                .addLast("http_encode", new ResponseEncoderHTTP1())
-                                .addLast("handle", new ChannelInboundHandlerAdapter() {
-                                    @Override
-                                    public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-                                        if (msg instanceof Request) {
-                                            Request request = (Request) msg;
-                                            System.out.printf("%s to \"%s\" from %s\n", request.type(), request.query(), ctx.channel().remoteAddress());
-                                            if (!"/".equals(request.query().toString()))   {
-                                                throw new GenericHTTPException(StatusCodes.Not_Found, request.query().toString());
-                                            }
-                                            ctx.channel().writeAndFlush(new Response.Simple(StatusCodes.OK, Unpooled.wrappedBuffer("ok".getBytes(StandardCharsets.US_ASCII)), Collections.emptyMap())).channel().close();
-                                        } else {
-                                            System.out.printf("[ERROR] Received invalid message (type: \"%s\"): %s\n", PorkUtil.className(msg), msg);
-                                        }
-                                    }
-                                })
-                                .addLast("http_exception", new ExceptionHandlerServerHTTP());
-                    }
-                })
-                .bind(8080).syncUninterruptibly().channel();
+public interface HttpServer {
+    /**
+     * @return the {@link HttpEngine} that created this {@link HttpServer}
+     */
+    HttpEngine engine();
 
-        new Scanner(System.in).nextLine();
+    /**
+     * Closes this server instance.
+     * <p>
+     * This will release any resources allocated specifically by this server (however, any resources allocated
+     * by the parent {@link HttpEngine} will not be released unless specifically stated otherwise by the
+     * implementation).
+     */
+    Future<Void> close();
 
-        ch.close().addListener(c -> System.exit(0));
-    }
+    /**
+     * @return a {@link Future} that will be notified when this {@link HttpServer} is closed
+     */
+    Future<Void> closeFuture();
 }
